requirements.txt
anthropic>=0.40.0
pdfplumber>=0.11.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
streamlit>=1.40.0
numpy>=1.26.0
requests>=2.32.0
tqdm>=4.66.0
python-dotenv>=1.0.0
scikit-learn>=1.5.0
